+++ database/truliaImageScraper/truliaHomeProfileUrls.js
@@ -1,0 +1,6 @@
+const truliaProfileElements = require('./truliaProfileElements');
+const getTruliaHomeProfileUrls = require('./getTruliaHomeProfileUrls');
+
+const truliaHomeProfileUrls = getTruliaHomeProfileUrls(truliaProfileElements);
+
+module.exports = truliaHomeProfileUrls;
