+++ database/truliaImageScraper/truliaProfileElements.js
@@ -1,0 +1,3 @@
+const truliaProfileElements = `<img id="photo-gallery-0" class="Picture__Image-sc-1exw3ow-0 cQtjRq GridGalleryImage__StyledPicture-sc-1jq3b2t-0 ikFQbA" src="https://www.trulia.com/pictures/thumbs_5/zillowstatic/fp/ecd66cd9b3c1b9f140d3862b3d0f0496-full.jpg" alt=""<img id="photo-gallery-0" class="Picture__Image-sc-1exw3ow-0 cQtjRq GridGalleryImage__StyledPicture-sc-1jq3b2t-0 ikFQbA" src="https://www.trulia.com/pictures/thumbs_5/zillowstatic/fp/1125c687625bb7f182fb0e25f8329473-full.jpg" alt=""><img id="photo-gallery-0" class="Picture__Image-sc-1exw3ow-0 cQtjRq GridGalleryImage__StyledPicture-sc-1jq3b2t-0 ikFQbA" src="https://www.trulia.com/pictures/thumbs_5/zillowstatic/fp/0d41509b105149e2202fc9b2ae4b2389-full.jpg" alt=""><img id="photo-gallery-0" class="Picture__Image-sc-1exw3ow-0 cQtjRq GridGalleryImage__StyledPicture-sc-1jq3b2t-0 ikFQbA" src="https://www.trulia.com/pictures/thumbs_5/zillowstatic/fp/572d2a60857bc722964ceda791942990-full.jpg" alt="">`;
+
+module.exports = truliaProfileElements;
